Add entities.py and utils.py
Made a PhysicsEntity class to update the position of the entities based on movement and velocity.
The rendering of the entity was also moved in this class.
In utils.py we use the path to load the image and also convert it to make it more efficient for rendering.
We import and use these in game.py